refactor: lib core-eval-support
